libnvpair: comparison of integers of different signs
../inet/rcmd.c:283:40: error: comparison of integers of different signs: 'ssize_t' (aka 'int') and 'unsigned int' [-Werror,-Wsign-compare]
                if (write(s, aport, strlen(aport)+1) != strlen(aport)+1) {
                    ~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~ ^  ~~~~~~~~~~~~~~~
1 error generated.

/code/illumos-gate/usr/src/common/nvpair/nvpair.c:733:21: error: comparison of integers of different signs: 'int32_t' (aka 'int') and 'unsigned long' [-Werror,-Wsign-compare]
            (nvp->nvp_size < NVP_SIZE_CALC(nvp->nvp_name_sz, 0)))
             ~~~~~~~~~~~~~ ^ ~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~
/code/illumos-gate/usr/src/common/nvpair/nvpair.c:740:32: error: comparison of integers of different signs: 'size_t' (aka 'unsigned int') and 'int' [-Werror,-Wsign-compare]
        return (strlen(NVP_NAME(nvp)) == nvp->nvp_name_sz - 1 ? 0 : EFAULT);
                ~~~~~~~~~~~~~~~~~~~~~ ^  ~~~~~~~~~~~~~~~~~~~~
/code/illumos-gate/usr/src/common/nvpair/nvpair.c:755:17: error: comparison of integers of different signs: 'int' and 'uint_t' (aka 'unsigned int') [-Werror,-Wsign-compare]
                for (i = 0; i < nelem; i++)
                            ~ ^ ~~~~~
/code/illumos-gate/usr/src/common/nvpair/nvpair.c:792:25: error: comparison of integers of different signs: 'int' and 'unsigned long' [-Werror,-Wsign-compare]
        if (size2 < 0 || size1 != NV_ALIGN(size2))
                         ~~~~~ ^  ~~~~~~~~~~~~~~~
/code/illumos-gate/usr/src/common/nvpair/nvpair.c:1177:14: error: comparison of integers of different signs: 'int' and 'unsigned long long' [-Werror,-Wsign-compare]
        if (name_sz >= 1ULL << (sizeof (nvp->nvp_name_sz) * NBBY - 1))
            ~~~~~~~ ^  ~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~
/code/illumos-gate/usr/src/common/nvpair/nvpair.c:1992:34: error: comparison of integers of different signs: 'size_t' (aka 'unsigned int') and 'int' [-Werror,-Wsign-compare]
                            (strlen(nvpair_name(nvp)) != n))
                             ~~~~~~~~~~~~~~~~~~~~~~~~ ^  ~
/code/illumos-gate/usr/src/common/nvpair/nvpair.c:2492:17: error: comparison of integers of different signs: 'int' and 'size_t' (aka 'unsigned int') [-Werror,-Wsign-compare]
                for (i = 0; i < nelem; i++)
                            ~ ^ ~~~~~
/code/illumos-gate/usr/src/common/nvpair/nvpair.c:2502:17: error: comparison of integers of different signs: 'int' and 'size_t' (aka 'unsigned int') [-Werror,-Wsign-compare]
                for (i = 0; i < nelem; i++) {
                            ~ ^ ~~~~~
/code/illumos-gate/usr/src/common/nvpair/nvpair.c:2515:17: error: comparison of integers of different signs: 'int' and 'size_t' (aka 'unsigned int') [-Werror,-Wsign-compare]
                for (i = 0; i < nelem; i++) {
                            ~ ^ ~~~~~
/code/illumos-gate/usr/src/common/nvpair/nvpair.c:2978:48: error: comparison of integers of different signs: 'unsigned long' and 'int32_t' (aka 'int') [-Werror,-Wsign-compare]
        if (NVP_SIZE_CALC(nvp->nvp_name_sz, value_sz) > nvp->nvp_size)
            ~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~ ^ ~~~~~~~~~~~~~
/code/illumos-gate/usr/src/common/nvpair/nvpair.c:3250:13: error: comparison of integers of different signs: 'uint_t' (aka 'unsigned int') and 'int' [-Werror,-Wsign-compare]
        if (buflen < value_sz)
            ~~~~~~ ^ ~~~~~~~~
/code/illumos-gate/usr/src/common/nvpair/nvpair.c:3363:17: error: comparison of integers of different signs: 'int' and 'uint_t' (aka 'unsigned int') [-Werror,-Wsign-compare]
                for (i = 0; i < nelem; i++) {
                            ~ ^ ~~~~~
12 errors generated.

../libnvpair.c:1005:11: error: comparison of integers of different signs: 'int' and 'uint_t' (aka 'unsigned int') [-Werror,-Wsign-compare]
                    (ai < a_len)) {
                     ~~ ^ ~~~~~
../libnvpair.c:1035:11: error: comparison of integers of different signs: 'int' and 'uint_t' (aka 'unsigned int') [-Werror,-Wsign-compare]
                    (ai < a_len) &&
                     ~~ ^ ~~~~~
../libnvpair.c:1058:11: error: comparison of integers of different signs: 'int' and 'uint_t' (aka 'unsigned int') [-Werror,-Wsign-compare]
                    (ai < a_len) &&
                     ~~ ^ ~~~~~
../libnvpair.c:1081:11: error: comparison of integers of different signs: 'int' and 'uint_t' (aka 'unsigned int') [-Werror,-Wsign-compare]
                    (ai < a_len) &&
                     ~~ ^ ~~~~~
../libnvpair.c:1104:11: error: comparison of integers of different signs: 'int' and 'uint_t' (aka 'unsigned int') [-Werror,-Wsign-compare]
                    (ai < a_len) &&
                     ~~ ^ ~~~~~
../libnvpair.c:1127:11: error: comparison of integers of different signs: 'int' and 'uint_t' (aka 'unsigned int') [-Werror,-Wsign-compare]
                    (ai < a_len) &&
                     ~~ ^ ~~~~~
../libnvpair.c:1150:11: error: comparison of integers of different signs: 'int' and 'uint_t' (aka 'unsigned int') [-Werror,-Wsign-compare]
                    (ai < a_len) &&
                     ~~ ^ ~~~~~
../libnvpair.c:1173:11: error: comparison of integers of different signs: 'int' and 'uint_t' (aka 'unsigned int') [-Werror,-Wsign-compare]
                    (ai < a_len) &&
                     ~~ ^ ~~~~~
../libnvpair.c:1196:11: error: comparison of integers of different signs: 'int' and 'uint_t' (aka 'unsigned int') [-Werror,-Wsign-compare]
                    (ai < a_len) &&
                     ~~ ^ ~~~~~
../libnvpair.c:1219:11: error: comparison of integers of different signs: 'int' and 'uint_t' (aka 'unsigned int') [-Werror,-Wsign-compare]
                    (ai < a_len) &&
                     ~~ ^ ~~~~~
../libnvpair.c:1232:12: error: comparison of integers of different signs: 'boolean_t' and 'int32_t' (aka 'int') [-Werror,-Wsign-compare]
                    (val == val_arg))
                     ~~~ ^  ~~~~~~~
../libnvpair.c:1246:22: error: comparison of integers of different signs: 'boolean_t' and 'int32_t' (aka 'int') [-Werror,-Wsign-compare]
                    (val_array[ai] == val_arg))
                     ~~~~~~~~~~~~~ ^  ~~~~~~~
../libnvpair.c:1245:11: error: comparison of integers of different signs: 'int' and 'uint_t' (aka 'unsigned int') [-Werror,-Wsign-compare]
                    (ai < a_len) &&
                     ~~ ^ ~~~~~
13 errors generated.

+++ usr/src/lib/libsocket/inet/rcmd.c
@@ -280,7 +280,8 @@ int rcmd_af(char **ahost, unsigned short rport, const char *locuser,
 			goto bad;
 		(void) listen(s2, 1);
 		(void) snprintf(aport, MAX_SHORTSTRLEN, "%d", lport);
-		if (write(s, aport, strlen(aport)+1) != strlen(aport)+1) {
+		if (write(s, aport, strlen(aport) + 1) !=
+		    (ssize_t)strlen(aport) + 1) {
 			perror(dgettext(TEXT_DOMAIN,
 			    "write: setting up stderr"));
 			(void) close(s2);
